Create immediateSmaller.py
GFG
Immediate Smaller Element
Given an integer array arr of size n. For each element in the array, check whether the right adjacent element (on the next immediate position) of the array is smaller. If next element is smaller, update the current index to that element. If not, then  -1.
Solution with linear time complexity.